// src/app/core/models/Contact.ts
import { NumberSymbol } from "@angular/common/src/i18n/locale_data_api";

export class Contact {

    ContactType: string;
    CompanyName: string;
    ContactName: string;
    Street: string;
    City: string;
    State: string;
    Zip: string;
    PhoneNo: string;
    Fax: string;
    Email: string;
    Website: string;
}